Backend/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.16.5",
    "bcrypt==4.0.1",
    "email-validator>=2.3.0",
    "fastapi>=0.118.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.43",
    "uvicorn[standard]>=0.37.0",
]
